# api-specification.yaml
openapi: 3.0.3
info:
  title: Mini Commerce MSA API
  description: |
    한국형 이커머스 플랫폼을 위한 마이크로서비스 아키텍처 API 명세서
    
    ## 아키텍처 개요
    - **Frontend**: React + TypeScript + Tailwind CSS
    - **API Gateway**: Node.js + Express (포트 8080)
    - **마이크로서비스**: Node.js, Python/FastAPI, Go, Java/Spring Boot
    - **데이터베이스**: PostgreSQL + Redis
    - **인프라**: Kubernetes + Istio Service Mesh
    
    ## 인증
    - JWT (JSON Web Token) 기반 인증
    - Authorization: Bearer {token}
    
    ## 서비스 포트
    - Auth Service: 3001
    - Catalog Service: 3002 (내부 8080)
    - Cart Service: 3003
    - Order Service: 3004 (내부 8080)
    - Inventory Service: 3005
    - Payment Service: 3006
    - Notification Service: 3007
    - Gateway: 8080
    
  version: 1.0.0
  contact:
    name: Mini Commerce Team
    email: dev@minicommerce.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.minicommerce.com
    description: Production server
  - url: https://staging-api.minicommerce.com
    description: Staging server
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/orders:
    get:
      tags:
        - Order
      summary: 주문 목록 조회
      description: 사용자의 주문 목록을 조회합니다.
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: 페이지 번호
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: 페이지당 항목 수
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          description: 주문 상태 필터
          schema:
            type: string
            enum: [pending, confirmed, shipping, delivered, cancelled]
      responses:
        '200':
          description: 주문 목록 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Order
      summary: 주문 생성
      description: 장바구니의 상품들로 주문을 생성합니다.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - items
                - shipping_address
              properties:
                items:
                  type: array
                  items:
                    type: object
                    required:
                      - product_id
                      - quantity
                    properties:
                      product_id:
                        type: integer
                      quantity:
                        type: integer
                      options:
                        type: object
                shipping_address:
                  $ref: '#/components/schemas/Address'
                payment_method:
                  type: string
                  enum: [card, bank_transfer, virtual_account]
                coupon_code:
                  type: string
                memo:
                  type: string
      responses:
        '201':
          description: 주문 생성 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/orders/{orderId}:
    get:
      tags:
        - Order
      summary: 주문 상세 조회
      description: 특정 주문의 상세 정보를 조회합니다.
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          description: 주문 ID
          schema:
            type: integer
      responses:
        '200':
          description: 주문 상세 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/OrderDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Order
      summary: 주문 상태 수정
      description: 주문 상태를 수정합니다. (관리자만 가능)
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          description: 주문 ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [pending, confirmed, shipping, delivered, cancelled]
                tracking_number:
                  type: string
                memo:
                  type: string
      responses:
        '200':
          description: 주문 상태 수정 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== Inventory Service ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "스마트폰 케이스"
        description:
          type: string
          example: "고품질 실리콘 케이스"
        price:
          type: number
          format: decimal
          example: 15000.00
        original_price:
          type: number
          format: decimal
          example: 20000.00
        discount_rate:
          type: number
          format: decimal
          example: 25.0
        category_id:
          type: integer
          example: 1
        category_name:
          type: string
          example: "전자제품"
        stock_quantity:
          type: integer
          example: 100
        images:
          type: array
          items:
            type: string
          example: ["https://example.com/image1.jpg", "https://example.com/image2.jpg"]
        tags:
          type: array
          items:
            type: string
          example: ["케이스", "스마트폰", "보호"]
        is_featured:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Order:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        order_number:
          type: string
          example: "ORD-20231201-001"
        status:
          type: string
          enum: [pending, confirmed, shipping, delivered, cancelled]
          example: "pending"
        total_amount:
          type: number
          format: decimal
          example: 45000.00
        shipping_address:
          $ref: '#/components/schemas/Address'
        payment_method:
          type: string
          example: "card"
        payment_status:
          type: string
          enum: [pending, paid, failed, refunded]
          example: "pending"
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OrderDetail:
      allOf:
        - $ref: '#/components/schemas/Order'
        - type: object
          properties:
            tracking_number:
              type: string
              example: "1234567890"
            shipping_info:
              type: object
              properties:
                carrier:
                  type: string
                  example: "CJ대한통운"
                estimated_delivery:
                  type: string
                  format: date-time
            payment_info:
              type: object
              properties:
                transaction_id:
                  type: string
                  example: "TXN-123456789"
                paid_at:
                  type: string
                  format: date-time
    OrderItem:
      type: object
      properties:
        id:
          type: integer
          example: 1
        order_id:
          type: integer
          example: 1
        product_id:
          type: integer
          example: 1
        product:
          $ref: '#/components/schemas/Product'
        quantity:
          type: integer
          example: 2
        price:
          type: number
          format: decimal
          example: 15000.00
        total_price:
          type: number
          format: decimal
          example: 30000.00
    Address:
      type: object
      properties:
        name:
          type: string
          example: "홍길동"
        phone:
          type: string
          example: "010-1234-5678"
        address_line1:
          type: string
          example: "서울시 강남구 테헤란로 123"
        address_line2:
          type: string
          example: "456호"
        city:
          type: string
          example: "서울시"
        state:
          type: string
          example: "강남구"
        postal_code:
          type: string
          example: "06292"
        country:
          type: string
          example: "대한민국"
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        total_pages:
          type: integer
          example: 5
        has_next:
          type: boolean
          example: true
        has_prev:
          type: boolean
          example: false
  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: "잘못된 요청입니다."
              errors:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                    message:
                      type: string
    Unauthorized:
      description: 인증 실패
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: "인증이 필요합니다."
...
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: "요청한 리소스를 찾을 수 없습니다."
...